@AkihiroSuda 2024-11-08
Thanks, do you have a benchmark result?