@nirs 2024-11-08
> Thanks, do you have a benchmark result?

I used both configurations when collecting data for #58 and I don't remember any difference. I can do another run on macOS 15 and M2 Max.