@nirs 2024-11-08
Rebased, I'll do quick test run to validate the lack of performance improvement and update the commit message.