@AkihiroSuda 2024-11-18
> @AkihiroSuda added test results, confirming that there is no improvement, and also showing that we have a performance regression in macOS 15.1 (not related to this change).

What is the benefit of this PR?