@nirs 2024-11-18
> What is the benefit of this PR?

- Avoiding unneeded queues for initialization and shutdown - we use the single serial host queue
- Avoiding the unwanted case when additional threads are created for serving vmnet traffic. Testing show that this rare enough so we don't see any effect on performance.
- Improve documentation and structure of the program

We can also use the serial queue later to simpilify locking - if we dispatch add/remove connection to the serial queue, no locking is needed. However this does not solve the need for locking when forwarding packets from one vms to other vms, since it is done in one of the vms queues.

I'm using this for few months without any issues.